agents/openai_sdk/src/agent/pyproject.toml
[project]
name = "OpenAISDKAgent"
version = "0.1.0"
description = "An agent that is built using the OpenAI SDK"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-adk[otel-gcp]",
    "litellm",
    "httpx",
    "python-dotenv",
    "uvicorn",
    "fastapi",
    "pydantic",
    "opentelemetry-api",
    "opentelemetry-sdk",
    "opentelemetry-exporter-otlp",
    "opentelemetry-instrumentation-httpx",
    "opentelemetry-instrumentation-system-metrics",
    "openinference-instrumentation-google-adk",
    "opentelemetry-instrumentation-threading",
    "openinference-instrumentation-mcp",
    "opentelemetry-instrumentation-fastapi",
    "openinference-instrumentation-litellm",
    "openinference-instrumentation-anthropic",
    "openinference-instrumentation-openai",
    "opentelemetry-instrumentation-aiohttp-client",
    "openai",
    "openai-agents"
]
